fix: deletable failed iou
